+++ backend/app/Domain/Report/Services/PromptExportServiceInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Report;
+namespace App\Domain\Report\Services;
 
 use App\Domain\Report\Models\Report;
