+++ backend/app/Http/Controllers/ReportController.php
@@ -19,8 +19,8 @@ use App\Domain\GitLab\Models\Commit;
 use App\Domain\Report\Models\Report;
 use App\Domain\Settings\Models\Setting;
 use App\Domain\Bitrix24\Models\Task;
-use App\Services\Report\PromptExportServiceInterface;
-use App\Services\Report\ReportExporterInterface;
+use App\Domain\Report\Services\PromptExportServiceInterface;
+use App\Domain\Report\Services\ReportExporterInterface;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
