+++ backend/app/Providers/ReportServiceProvider.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Providers;
 
 use App\Services\Report\PromptExportService;
-use App\Services\Report\PromptExportServiceInterface;
-use App\Services\Report\ReportExporterInterface;
+use App\Domain\Report\Services\PromptExportServiceInterface;
+use App\Domain\Report\Services\ReportExporterInterface;
 use App\Services\Report\WordExporter;
 use Illuminate\Support\ServiceProvider;
